requirements.txt
Flask
eth-account
requests
python-binance